refactor(normalizer): merge rules.py into merchant_normalizer.py
The rule-repository loader (ChannelRule / RuleSet / load_rules /
build_system_prompt) is only used by the normalizer, so fold it into
merchant_normalizer.py and drop the separate rules.py module. Importers
updated to pull these from the package (or merchant_normalizer directly).
